Saco warninigs

+++ src/main/gfx/health_bar.cpp
@@ -26,7 +26,7 @@ void HealthBar::visit(Unit& entity) {
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaFondo);
 	double porcentaje = (double)entity.health.get() / (double)entity.health.max;
 	SDL_SetRenderDrawColor(owner.owner.getRenderer(), 255, 0, 0, 255);
-	SDL_Rect lineaSalud = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 2), TILE_WIDTH_DEFAULT / 2 * porcentaje, 2 };
+	SDL_Rect lineaSalud = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 2), (int)(TILE_WIDTH_DEFAULT / 2 * porcentaje), 2 };
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaSalud);
 }
 
@@ -38,7 +38,7 @@ void HealthBar::visit(Building& entity) {
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaFondo);
 	double porcentaje = (double)entity.health.get() / (double)entity.health.max;
 	SDL_SetRenderDrawColor(owner.owner.getRenderer(), 255, 0, 0, 255);
-	SDL_Rect lineaSalud = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 2), TILE_WIDTH_DEFAULT / 2 * porcentaje, 2 };
+	SDL_Rect lineaSalud = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 2), (int)(TILE_WIDTH_DEFAULT / 2 * porcentaje), 2 };
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaSalud);
 	if (entity.isInAction) {
 		SDL_SetRenderDrawColor(owner.owner.getRenderer(), 0, 0, 0, 255);
@@ -46,7 +46,7 @@ void HealthBar::visit(Building& entity) {
 		SDL_RenderFillRect(owner.owner.getRenderer(), &lineaFondoProgress);
 		double porcentajeProgress = (double)entity.progress.get() / (double)entity.progress.max;
 		SDL_SetRenderDrawColor(owner.owner.getRenderer(), 0, 0, 255, 255);
-		SDL_Rect lineaProgreso = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 4), TILE_WIDTH_DEFAULT / 2 * porcentajeProgress, 2 };
+		SDL_Rect lineaProgreso = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 4), int(TILE_WIDTH_DEFAULT / 2 * porcentajeProgress), 2 };
 		SDL_RenderFillRect(owner.owner.getRenderer(), &lineaProgreso);
 	}
 }
@@ -59,7 +59,7 @@ void HealthBar::visit(UnfinishedBuilding& entity) {
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaFondo);
 	double porcentaje = (double)entity.progress.get() / (double)entity.progress.max;
 	SDL_SetRenderDrawColor(owner.owner.getRenderer(), 0, 0, 255, 255);
-	SDL_Rect lineaProgreso = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 4), TILE_WIDTH_DEFAULT / 2 * porcentaje, 2 };
+	SDL_Rect lineaProgreso = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 4), (int)(TILE_WIDTH_DEFAULT / 2 * porcentaje), 2 };
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaProgreso);
 }
 
@@ -71,7 +71,7 @@ void HealthBar::visit(Flag& entity) {
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaFondo);
 	double porcentaje = (double)entity.health.get() / (double)entity.health.max;
 	SDL_SetRenderDrawColor(owner.owner.getRenderer(), 255, 0, 0, 255);
-	SDL_Rect lineaSalud = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 2), TILE_WIDTH_DEFAULT / 2 * porcentaje, 2 };
+	SDL_Rect lineaSalud = { (int)((centro.x - TILE_WIDTH_DEFAULT / 4)), (int)(centro.y + 2), (int)(TILE_WIDTH_DEFAULT / 2 * porcentaje), 2 };
 	SDL_RenderFillRect(owner.owner.getRenderer(), &lineaSalud);
 }
